Вноси правки в открытие модалок
