+++ js/upload-photo-form.js
@@ -1,34 +1,39 @@
-import {isEscapeKey} from './util.js';
+// import {isEscapeKey} from './util.js';
 
-const body = document.querySelector('body');
+// const body = document.querySelector('body');
 
-const uploadForm = body.querySelector('.img-upload__form');
-const uploadInput = uploadForm.querySelector('.img-upload__input');
-const uploadFormEdit = uploadForm.querySelector('.img-upload__overlay');
-const uploadCloseButton = uploadForm.querySelector('.img-upload__cancel');
+// const uploadForm = body.querySelector('.img-upload__form');
+// const uploadInput = uploadForm.querySelector('.img-upload__input');
+// const uploadFormEdit = uploadForm.querySelector('.img-upload__overlay');
+// const uploadCloseButton = uploadForm.querySelector('.img-upload__cancel');
+// const hashtagText = uploadForm.querySelector('.text__hashtags');
+// const descriptionText = uploadForm.querySelector('.text__description');
 
-const onDocumentKeydownForm = (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.preventDefault();
-    closeFormModal();
-  }
-};
+// const onDocumentKeydownForm = (evt) => {
+//   if (isEscapeKey(evt)) {
+//     evt.preventDefault();
+//     closeFormModal();
+//   }
+// };
 
-function openFormModal () {
-  uploadFormEdit.classList.remove('hidden');
-  body.classList.add('modal-open');
-  document.addEventListener('keydown', onDocumentKeydownForm);
-}
+// function openFormModal () {
+//   uploadFormEdit.classList.remove('hidden');
+//   body.classList.add('modal-open');
+//   document.addEventListener('keydown', onDocumentKeydownForm);
+// }
 
-function closeFormModal () {
-  uploadFormEdit.classList.add('hidden');
-  body.classList.remove('modal-open');
-  document.removeEventListener('keydown', onDocumentKeydownForm);
-  uploadInput.value = '';
-}
+// function closeFormModal () {
+//   uploadFormEdit.classList.add('hidden');
+//   body.classList.remove('modal-open');
+//   document.removeEventListener('keydown', onDocumentKeydownForm);
 
-uploadInput.addEventListener('click', openFormModal);
+//   uploadInput.value = '';
+//   hashtagText.value = '';
+//   descriptionText.value = '';
+// }
 
-uploadCloseButton.addEventListener('click', closeFormModal);
+// uploadInput.addEventListener('click', openFormModal);
 
-export {openFormModal};
+// uploadCloseButton.addEventListener('click', closeFormModal);
+
+// export {openFormModal};
